+++ src/Components/Icon/index.tsx
@@ -1,0 +1,1 @@
+export * from './Icon'
