+++ src/Components/Label/index.tsx
@@ -1,0 +1,1 @@
+export * from './Label'
