+++ src/index.ts
@@ -39,7 +39,7 @@ export * from './Components/Form/FormHelpText'
 export * from './Components/Form/FormLabel'
 export * from './Components/Form/FormValidationElement'
 export * from './Components/Grid'
-export * from './Components/Icon/Icon'
+export * from './Components/Icon'
 export * from './Components/IndexList/IndexList'
 export * from './Components/IndexList/IndexListBody'
 export * from './Components/IndexList/IndexListHeader'
@@ -50,7 +50,7 @@ export * from './Components/Inputs/Composed/AutoInput'
 export * from './Components/Inputs/Composed/RangeSlider'
 export * from './Components/Inputs/Input'
 export * from './Components/Inputs/TextArea'
-export * from './Components/Label/Label'
+export * from './Components/Label'
 export * from './Components/NavMenu/NavMenu'
 export * from './Components/NavMenu/NavMenuItem'
 export * from './Components/NavMenu/NavMenuSubItem'
@@ -84,10 +84,8 @@ export * from './Components/ResourceList/List/ResourceList'
 export * from './Components/ResourceList/List/ResourceListBody'
 export * from './Components/ResourceList/List/ResourceListHeader'
 export * from './Components/ResourceList/List/ResourceListSorter'
-export * from './Components/SelectableCard/SelectableCard'
+export * from './Components/SelectableCard'
 export * from './Components/SlideToggle'
-export * from './Components/SlideToggle/SlideToggle'
-export * from './Components/SlideToggle/SlideToggleLabel'
 export * from './Components/Spinners/SparkleSpinner'
 export * from './Components/Spinners/SpinnerContainer'
 export * from './Components/Spinners/TechnoSpinner'
